Two birds one stone bug fix
As a player, I do not want to see a floating health bar when I have walked to the next room so that it actually looks like I left the room.

As a player, I don't want to see a floating weapon animation as I am leaving the room so that it seems I have actually left the room.